# backend/src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "SNM Medical API",
    "description": "API documentation for SNM Medical application",
    "version": "1.0.0",
    "contact": {
      "name": "SNM Medical Team",
      "email": "support@snmmedical.com"
    }
  },
  "host": "localhost:5000",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your bearer token in the format \"Bearer {token}\""
    }
  },
  "consumes": [
    "application/json",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/profile": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get user profile",
        "description": "Retrieve the profile of the currently authenticated user.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Update user profile",
        "description": "Update the profile of the currently authenticated user.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "type": "string",
                  "example": "John"
                },
                "lastName": {
                  "type": "string",
                  "example": "Doe"
                },
                "mobileNumber": {
                  "type": "string",
                  "example": "9876543210"
                },
                "address": {
                  "type": "string",
                  "example": "123 Main Street"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/registration": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Get registration reports",
        "description": "Retrieve registration reports data. (Implementation pending)",
        "responses": {
          "200": {
            "description": "Registration reports retrieved successfully"
          },
          "500": {
            "description": "Registration reports failed"
          }
        }
      }
    }
  },
...
}